[P3b-T01] commit 2: relocate fe-common security.authentication.* (13 classes) to fe-kerberos
Relocate the Hadoop authentication mechanism from fe-common package
org.apache.doris.common.security.authentication to the neutral leaf module
fe-kerberos (org.apache.doris.kerberos), making fe-kerberos the single source
of truth (D-017 / P3b commit 2 of 3).

Root: the auth machinery lived in fe-common, so connectors that cannot depend
on fe-common were cut off from it, and it was duplicated by a fe-kerberos facts
slice (AuthType). This collapses the fe-common copy into fe-kerberos.

Moved (git mv, history-preserving, package line changed): AuthenticationConfig,
ExecutionAuthenticator, HadoopAuthenticator, HadoopExecutionAuthenticator,
HadoopKerberosAuthenticator, HadoopSimpleAuthenticator,
ImpersonatingHadoopAuthenticator, KerberosAuthenticationConfig,
KerberosTicketUtils, PreExecutionAuthenticator, PreExecutionAuthenticatorCache,
SimpleAuthenticationConfig (+ moved tests AuthenticationTest,
KerberosTicketUtilsTest).

AuthType merge: the fe-common AuthType (mutable getCode/setCode/setDesc - all
dead, zero callers) is folded into the existing immutable fe-kerberos AuthType;
added isSupportedAuthType(String) (sole caller HiveTable), kept getDesc/fromString.
fe-common AuthType deleted.

Retargeted 41 consumer imports (24 fe-core main + 14 fe-core test + 3
be-java-extensions JNI scanners) common.security.authentication.* -> kerberos.*
(import-line only, no logic change - incl. the otherwise-forbidden
datasource.property.{storage,metastore} packages under the D-017 import-only
exception), with import blocks re-sorted to satisfy CustomImportOrder.

pom: fe-kerberos gains hadoop-common(provided)+guava+commons-lang3+lombok+
log4j-api (mirroring fe-common; leaf invariant intact - zero FE module deps);
fe-common gains a compile edge to fe-kerberos (re-provides the relocated classes
transitively to fe-core et al.); java-common gains an explicit fe-kerberos edge
for the BE-java scanner plugins.

Leaf invariant confirmed: none of the 13 classes import any non-kerberos doris
type (recon wf_d5566c5f-7b1, adversarially verified); build order already correct
(fe-kerberos before fe-common).

Verify: fe-kerberos 11/0 (incl. moved AuthenticationTest + KerberosTicketUtilsTest);
fe-core + java-common + 3 scanners test-compile BUILD SUCCESS; checkstyle 0;
connector import-gate exit 0; whole-repo grep common.security.authentication = 0.
docker kerberos e2e NOT run (the real gate; lands after commit 3 unifies the two
HadoopAuthenticator interfaces + deletes the fe-filesystem-hdfs copies).
