+++ fe/be-java-extensions/java-common/pom.xml
@@ -36,6 +36,11 @@ under the License.
     <dependencies>
         <dependency>
             <groupId>org.apache.doris</groupId>
+            <artifactId>fe-kerberos</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.doris</groupId>
             <artifactId>fe-common</artifactId>
             <version>${project.version}</version>
         </dependency>
