+++ fe/fe-core/src/main/java/org/apache/doris/catalog/HdfsResource.java
@@ -19,7 +19,7 @@ package org.apache.doris.catalog;
 
 import org.apache.doris.common.DdlException;
 import org.apache.doris.common.proc.BaseProcResult;
-import org.apache.doris.common.security.authentication.AuthenticationConfig;
+import org.apache.doris.kerberos.AuthenticationConfig;
 import org.apache.doris.thrift.THdfsConf;
 import org.apache.doris.thrift.THdfsParams;
