+++ fe/fe-core/src/main/java/org/apache/doris/datasource/property/metastore/AbstractPaimonProperties.java
@@ -17,11 +17,11 @@
 
 package org.apache.doris.datasource.property.metastore;
 
-import org.apache.doris.common.security.authentication.ExecutionAuthenticator;
-import org.apache.doris.common.security.authentication.HadoopExecutionAuthenticator;
 import org.apache.doris.datasource.property.storage.HdfsProperties;
 import org.apache.doris.datasource.property.storage.StorageProperties;
 import org.apache.doris.foundation.property.ConnectorProperty;
+import org.apache.doris.kerberos.ExecutionAuthenticator;
+import org.apache.doris.kerberos.HadoopExecutionAuthenticator;
 
 import lombok.Getter;
