+++ fe/fe-core/src/main/java/org/apache/doris/datasource/property/metastore/HMSBaseProperties.java
@@ -19,12 +19,12 @@ package org.apache.doris.datasource.property.metastore;
 
 import org.apache.doris.common.CatalogConfigFileUtils;
 import org.apache.doris.common.Config;
-import org.apache.doris.common.security.authentication.AuthenticationConfig;
-import org.apache.doris.common.security.authentication.HadoopAuthenticator;
-import org.apache.doris.common.security.authentication.KerberosAuthenticationConfig;
 import org.apache.doris.foundation.property.ConnectorPropertiesUtils;
 import org.apache.doris.foundation.property.ConnectorProperty;
 import org.apache.doris.foundation.property.ParamRules;
+import org.apache.doris.kerberos.AuthenticationConfig;
+import org.apache.doris.kerberos.HadoopAuthenticator;
+import org.apache.doris.kerberos.KerberosAuthenticationConfig;
 
 import com.google.common.base.Strings;
 import lombok.Getter;
