+++ fe/fe-core/src/main/java/org/apache/doris/datasource/property/metastore/HiveHMSProperties.java
@@ -18,8 +18,8 @@
 package org.apache.doris.datasource.property.metastore;
 
 import org.apache.doris.common.Config;
-import org.apache.doris.common.security.authentication.HadoopExecutionAuthenticator;
 import org.apache.doris.foundation.property.ConnectorProperty;
+import org.apache.doris.kerberos.HadoopExecutionAuthenticator;
 
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
