+++ fe/fe-core/src/main/java/org/apache/doris/datasource/property/metastore/PaimonHMSMetaStoreProperties.java
@@ -17,8 +17,8 @@
 
 package org.apache.doris.datasource.property.metastore;
 
-import org.apache.doris.common.security.authentication.HadoopExecutionAuthenticator;
 import org.apache.doris.foundation.property.ConnectorProperty;
+import org.apache.doris.kerberos.HadoopExecutionAuthenticator;
 
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
