+++ fe/fe-core/src/main/java/org/apache/doris/datasource/property/storage/HdfsCompatibleProperties.java
@@ -17,9 +17,9 @@
 
 package org.apache.doris.datasource.property.storage;
 
-import org.apache.doris.common.security.authentication.HadoopAuthenticator;
-import org.apache.doris.common.security.authentication.HadoopSimpleAuthenticator;
-import org.apache.doris.common.security.authentication.SimpleAuthenticationConfig;
+import org.apache.doris.kerberos.HadoopAuthenticator;
+import org.apache.doris.kerberos.HadoopSimpleAuthenticator;
+import org.apache.doris.kerberos.SimpleAuthenticationConfig;
 
 import com.google.common.collect.ImmutableSet;
 import lombok.Getter;
