+++ fe/fe-core/src/main/java/org/apache/doris/datasource/property/storage/HdfsProperties.java
@@ -18,8 +18,8 @@
 package org.apache.doris.datasource.property.storage;
 
 import org.apache.doris.common.UserException;
-import org.apache.doris.common.security.authentication.HadoopAuthenticator;
 import org.apache.doris.foundation.property.ConnectorProperty;
+import org.apache.doris.kerberos.HadoopAuthenticator;
 
 import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableSet;
