+++ fe/fe-core/src/test/java/org/apache/doris/datasource/hive/HmsCommitTest.java
@@ -20,12 +20,12 @@ package org.apache.doris.datasource.hive;
 import org.apache.doris.analysis.UserIdentity;
 import org.apache.doris.catalog.Column;
 import org.apache.doris.catalog.PrimitiveType;
-import org.apache.doris.common.security.authentication.ExecutionAuthenticator;
 import org.apache.doris.common.util.DebugUtil;
 import org.apache.doris.datasource.NameMapping;
 import org.apache.doris.datasource.TestHMSCachedClient;
 import org.apache.doris.filesystem.local.LocalFileSystem;
 import org.apache.doris.fs.SpiSwitchingFileSystem;
+import org.apache.doris.kerberos.ExecutionAuthenticator;
 import org.apache.doris.nereids.trees.plans.commands.insert.HiveInsertCommandContext;
 import org.apache.doris.qe.ConnectContext;
 import org.apache.doris.thrift.THiveLocationParams;
