+++ fe/fe-kerberos/src/main/java/org/apache/doris/kerberos/ExecutionAuthenticator.java
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package org.apache.doris.common.security.authentication;
+package org.apache.doris.kerberos;
 
 import java.util.concurrent.Callable;
